[1.0.0] handle render req contact sent to view, fix login facebok and login google
