+++ src/controller/index.js
@@ -1,7 +1,9 @@
 import authController from './authController';
-import homeController from './homeController';
+import {homeController} from './homeController';
 import contactController from './contactController';
+import userController from './userController';
 
 module.exports.auth = authController;
 module.exports.home = homeController;
 module.exports.contact = contactController;
+module.exports.user = userController;
