+++ src/public/javascript/mainConfig.js
@@ -3,4 +3,5 @@ let loadingModal = $(".loading-modal");
 $(document).ready(function(){
   loadingModal.hide();
   
+ 
 })
