+++ src/services/index.js
@@ -1,7 +1,9 @@
 import authServices from './authServices';
 import homeServices from './homeServices';
 import contactService from './contactService';
+import userServices from './userServices';
 
 module.exports.auth = authServices;
 module.exports.home = homeServices;
 module.exports.contact = contactService;
+module.exports.user = userServices;
